installations/test_py_installations.py
import os
import sys
import ast
import collections
import copy
import glob
import yaml
import time

import torch
import torch.nn as nn
import torch.nn.functional as F
from torchvision.transforms import Compose

import numpy as np
from PIL import Image
import PIL.ImageColor as ImageColor
import PIL.ImageDraw as ImageDraw
import PIL.ImageFont as ImageFont
import six

sys.path.append('/opt/ros/kinetic/lib/python2.7/dist-packages')
sys.path.remove('/opt/ros/kinetic/lib/python2.7/dist-packages')
import cv2


print('')
print('[INFO] Python Installations Successfull')

